further cleaning of environment.rs
